+++ frontend/src/pages/publik/kamus/index.js
@@ -1,6 +1,7 @@
 export { default as Ejaan } from './Ejaan';
 export { default as Gramatika } from './Gramatika';
 export { default as Kamus } from './Kamus';
+export { default as KamusAcak } from './KamusAcak';
 export { default as KamusDetail } from './KamusDetail';
 export { default as Makna } from './Makna';
 export { default as Rima } from './Rima';